+++ manualTests/mockRoomImportEnforcer.ts
@@ -1,0 +1,7 @@
+import { iRoomImportEnforcer } from '../src';
+
+export class MockRoomImportEnforcer implements iRoomImportEnforcer {
+  public addRoomConstructor(): void {
+    // Nothing
+  }
+}